Create proper type tensors in padding batch
Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, kamil.andrzejewski

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, kamil.andrzejewski

JIRA Issues: AFS-23

Differential Revision: https://phabricator.sourcevertex.net/D81117

+++ poptorch_geometric/python/collate.py
@@ -92,7 +92,7 @@ def _reset_attr(value: Any, key: str = None) -> Any:
         # NOTE: It has to be torch.zeros - creating a Tensor directly
         # (through torch.tensor) with 0 in shape ends up in creating a
         # tensor with wrong dimensions.
-        return torch.zeros(_reset_dim(value.shape, key))
+        return torch.zeros(_reset_dim(value.shape, key), dtype=value.dtype)
     return type(value)()
